\name{Bookdown Template}
\alias{FFTemplates}
\title{Bookdown Template}
\usage{
File -> New File -> RMarkdown -> From Template -> Bookdown Template
}
\description{
This package provides a bookdown template to be used by the
Forsyth Futures' staff. The bibliography.bib file will need to be edited 
depending on whether the packages included are used, or if additional packages 
are leveraged.
}
