\name{Data Request Template}
\alias{Data Request Template}
\title{Data Request Template}
\usage{
File -> New File -> RMarkdown -> From Template -> Data Request Template
}
\description{
This package provides a data request template to be used by the
Forsyth Futures' staff. The preamble.tex file needs to be updated
with a title and organization name on lines 56 and 25, respectively.
The bibliography.bib file will need to be edited depending on whether
the packages included are used, or if additional packages are
leveraged.
}
